# Optimize startup time using local & lazy imports (take 2)