--- a/tests/test_backend.py
+++ b/tests/test_backend.py
@@ -4,6 +4,7 @@ import pytest
 import annif
 import annif.backend
 import annif.corpus
+import importlib.util
 
 
 def test_get_backend_nonexistent():
@@ -50,3 +51,35 @@ def test_fill_params_with_defaults(project):
                        project=project)
     expected_default_params = {'limit': 100}
     assert expected_default_params == dummy.params
+
+
+@pytest.mark.skipif(importlib.util.find_spec("fasttext") is not None,
+                    reason="test requires that fastText is NOT installed")
+def test_get_backend_fasttext_not_installed():
+    with pytest.raises(ValueError) as excinfo:
+        annif.backend.get_backend('fasttext')
+    assert 'fastText not available' in str(excinfo.value)
+
+
+@pytest.mark.skipif(importlib.util.find_spec("tensorflow") is not None,
+                    reason="test requires that TensorFlow is NOT installed")
+def test_get_backend_nn_ensemble_not_installed():
+    with pytest.raises(ValueError) as excinfo:
+        annif.backend.get_backend('nn_ensemble')
+    assert 'TensorFlow not available' in str(excinfo.value)
+
+
+@pytest.mark.skipif(importlib.util.find_spec("omikuji") is not None,
+                    reason="test requires that Omikuji is NOT installed")
+def test_get_backend_omikuji_not_installed():
+    with pytest.raises(ValueError) as excinfo:
+        annif.backend.get_backend('omikuji')
+    assert 'Omikuji not available' in str(excinfo.value)
+
+
+@pytest.mark.skipif(importlib.util.find_spec("yake") is not None,
+                    reason="test requires that YAKE is NOT installed")
+def test_get_backend_yake_not_installed():
+    with pytest.raises(ValueError) as excinfo:
+        annif.backend.get_backend('yake')
+    assert 'YAKE not available' in str(excinfo.value)